@@ -8,7 +8,9 @@
     "build": "vite build",
     "preview": "vite preview",
     "deploy": "yarn run build && gh-pages -d dist",
-    "lint": "eslint ./src/ --ext .ts,.tsx"
+    "check-types": "tsc --noemit",
+    "lint:ts": "eslint . --ext .ts,.tsx --fix",
+    "lint": "yarn check-types && yarn lint:ts"
   },
   "repository": {
     "type": "git",
@@ -22,33 +24,35 @@
   },
   "homepage": "https://github.com/hsimpson/ts-raytracer#readme",
   "devDependencies": {
-    "@types/node": "^17.0.8",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
+    "@types/node": "^20.2.5",
+    "@types/react": "^18.2.7",
+    "@types/react-dom": "^18.2.4",
     "@types/recoil": "0.0.9",
-    "@typescript-eslint/eslint-plugin": "^5.9.0",
-    "@typescript-eslint/parser": "^5.9.0",
-    "@typescript-eslint/typescript-estree": "^5.9.0",
-    "@vitejs/plugin-react": "^1.1.4",
-    "@webgpu/types": "^0.1.9",
-    "eslint": "^8.6.0",
-    "eslint-config-prettier": "^8.3.0",
+    "@typescript-eslint/eslint-plugin": "^5.59.8",
+    "@typescript-eslint/parser": "^5.59.8",
+    "@typescript-eslint/typescript-estree": "^5.59.8",
+    "@vitejs/plugin-react": "^4.0.0",
+    "@webgpu/types": "^0.1.32",
+    "eslint": "^8.41.0",
+    "eslint-config-prettier": "^8.8.0",
     "eslint-config-react": "^1.1.7",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.28.0",
-    "eslint-plugin-react-hooks": "^4.3.0",
-    "gh-pages": "^3.2.3",
-    "prettier": "^2.5.1",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "gh-pages": "^5.0.0",
+    "prettier": "^2.8.8",
     "resize-observer-polyfill": "^1.5.1",
-    "rimraf": "^3.0.2",
-    "typescript": "^4.5.4",
-    "vite": "^2.7.10"
+    "rimraf": "^5.0.1",
+    "typescript": "^5.0.4",
+    "vite": "^4.3.9"
   },
   "dependencies": {
+    "@types/dom-webcodecs": "^0.1.7",
+    "alea": "^1.0.1",
     "gl-matrix": "^3.4.3",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "recoil": "0.5.2",
-    "simplex-noise": "^3.0.1"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "recoil": "0.7.7",
+    "simplex-noise": "^4.0.1"
   }
 }
